+++ NEWS.md
@@ -1,3 +1,8 @@
+# ggblanket 1.3.5
+
+* Reverted to scales::oob_keep 
+* Breaking: removed option to adjust oob. 
+
 # ggblanket 1.3.4
 
 * Added `facet_space` argument.